requirement.txt
numpy 
pandas
python-dotenv
psycopg2-binary
scikit-learn
seaborn
matplotlib
catboost
xgboost
mlflow
#-e .

